refactor(managers): reorganized AIProvider, AIContext and AITool managers
