+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Code cleanup in AI Chat Component, WebChatDialog and WebChatUtils
 - Renamed SmartHopper.Config to SmartHopper.Infrastructure
 - Renamed SmartHopper.Config.Tests to SmartHopper.Infrastructure.Tests
+- Reorganized AIProvider, AIContext and AITool managers
 
 ## [0.3.3-alpha] - 2025-06-23
