+++ src/SmartHopper.Core/AI/AIContextProviders.cs
@@ -11,7 +11,8 @@
 using System;
 using System.Collections.Generic;
 using Rhino;
-using SmartHopper.Infrastructure.Managers;
+using SmartHopper.Infrastructure.Interfaces;
+using SmartHopper.Infrastructure.Managers.AIContext;
 
 namespace SmartHopper.Core.AI
 {
