+++ src/SmartHopper.Infrastructure/Managers/AIProviders/AIProviderSettings.cs
@@ -1,0 +1,34 @@
+/*
+ * SmartHopper - AI-powered Grasshopper Plugin
+ * Copyright (C) 2025 Marc Roca Musach
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ */
+
+using System;
+using System.Collections.Generic;
+using SmartHopper.Infrastructure.Interfaces;
+using SmartHopper.Infrastructure.Models;
+
+namespace SmartHopper.Infrastructure.Managers.AIProviders
+{
+    /// <summary>
+    /// Base class for provider settings, encapsulating common UI building and persistence logic.
+    /// </summary>
+    public abstract class AIProviderSettings : IAIProviderSettings
+    {
+        protected readonly IAIProvider provider;
+
+        protected AIProviderSettings(IAIProvider provider)
+        {
+            this.provider = provider ?? throw new ArgumentNullException(nameof(provider));
+        }
+
+        public abstract IEnumerable<SettingDescriptor> GetSettingDescriptors();
+
+        public abstract bool ValidateSettings(Dictionary<string, object> settings);
+    }
+}
