+++ src/SmartHopper.Infrastructure/Managers/AITools/AIToolManager.cs
@@ -19,7 +19,7 @@ using Newtonsoft.Json.Linq;
 using SmartHopper.Infrastructure.Interfaces;
 using SmartHopper.Infrastructure.Models;
 
-namespace SmartHopper.Infrastructure.Managers
+namespace SmartHopper.Infrastructure.Managers.AITools
 {
     /// <summary>
     /// Central manager for AI tools that can be called from chat interfaces.
