+++ src/SmartHopper.Providers.MistralAI/MistralAIProviderSettings.cs
@@ -13,6 +13,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using SmartHopper.Infrastructure.Dialogs;
 using SmartHopper.Infrastructure.Interfaces;
+using SmartHopper.Infrastructure.Managers.AIProviders;
 using SmartHopper.Infrastructure.Models;
 
 namespace SmartHopper.Providers.MistralAI
