+++ src/SmartHopper.Providers.Template/TemplateProviderSettings.cs
@@ -8,12 +8,13 @@
  * version 3 of the License, or (at your option) any later version.
  */
 
-using SmartHopper.Infrastructure.Interfaces;
-using SmartHopper.Infrastructure.Managers;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using SmartHopper.Infrastructure.Dialogs;
+using SmartHopper.Infrastructure.Interfaces;
+using SmartHopper.Infrastructure.Managers.AIProviders;
+using SmartHopper.Infrastructure.Models;
 
 namespace SmartHopper.Providers.Template
 {
